Integrating Text Context with the ASP.NET Core Dependency Injection System
1. Build EmployeeServiceWithAspNetCoreDIFixture class.
a. Implement IDisposable

b. Instantiate ServiceCollection on the constructor, then AddScoped EmployeeFactory,EmployeeManagementTestDataRepository, and EmployeeService.

c. Call BuildServiceProvider on the constructor and store it on a readonly ServiceProvider field.

d. Expose EmployeeManagementTestDataRepository and EmployeeService using a getter that returns an IEmployeeManagementRepository and IEmployeeService.

2. Build EmployeeServiceTestsWithAspNetCoreDI
a. Implement IClassFixture passing in EmployeeServiceWithAspNetCoreDIFixture.

b. Inject the EmployeeServiceWithAspNetCoreDIFixture on the constructor and store it in a readonly property.

c. Implement the tests.